+++ lib/index.ts
@@ -1,4 +1,6 @@
 export { ALL_CONSUMERS } from './sqs.constants';
+export type { MessageSchema, Payload, SqsQueue } from './sqs.contract';
+export { defineQueue } from './sqs.contract';
 export * from './sqs.decorators';
 export * from './sqs.module';
 export * from './sqs.service';